% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_test.R
\name{get_mapply_predict_test}
\alias{get_mapply_predict_test}
\title{model predictions}
\usage{
get_mapply_predict_test(study = NULL, dataset = NULL, models = NULL,
  classvariable = NULL, classtype = NULL)
}
\arguments{
\item{study}{character, the names of the specified test dataset.}

\item{dataset}{list, a list contained multi test datasets.}

\item{models}{list, a list contained multi model object for predictions.}

\item{classvariable}{character, the header name of the train of test dataframe 
for the classification.}

\item{classtype}{character, the name of the positive group for classification.}
}
\description{
predict the multi-test datasets with the results of the multi-models.
}
\details{
TODO
}
\examples{


}
\author{
Shuangbin Xu
}
