% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_train_test_data.R
\name{get_train_test_data}
\alias{get_train_test_data}
\title{get train and test data}
\usage{
get_train_test_data(test_study, datasets)
}
\arguments{
\item{test_study}{character, the names of the list datasets.}

\item{datasets}{list, a list contianed the multi-dataframes with features and target infromation.}
}
\value{
a list contianed the test datasets and train datasets.
}
\description{
Build the train and test datasets with a multi-datasets.
}
\details{
TODO
}
\examples{



}
\author{
Shuangbin Xu
}
