% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getthresholds.R
\name{getthresholds}
\alias{getthresholds}
\title{Creat the thresholds vector.}
\usage{
getthresholds(dataset = NULL, var_of_interest = NULL,
  type = "median")
}
\arguments{
\item{dataset}{dataframe, a dataframe contain interesting variable.}

\item{var_of_interest}{character, a vector interesting variables}

\item{type}{character, the method of choose thresholds, (median or mean), default is median}
}
\value{
a vector thresholds of interesting variable.
}
\description{
Creat the thresholds for the \code{\link[MetaMicrobiome]{high_low_vector}}
}
\details{
TODO
}
\examples{

library("MetaMicrobiome")
data <- read.csv(system.file("data", 
         package="MetaMicrobiome", 
      "Baxter_16_alpha_data.csv")) 
thresVetor <- getthresholds(dataset=data, 
              var_of_interest=c("Shannon", 
                                "Observe"), 
              type="median")

}
\author{
Shuangbin Xu
}
