% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest.R
\name{ggforest}
\alias{ggforest}
\title{forest plot base ggplot2 with the result of RunPoolEffect and multiVarRRTab.}
\usage{
ggforest(dataset = NULL, manualcolors = NULL, manualshapes = NULL,
  Logscale = TRUE, x = NULL, y = NULL, lower = NULL,
  upper = NULL, pointsize = 2, linesize = 0.4,
  errorbarhheight = 0.2, colorVar = NULL, shapeVar = NULL,
  facetx = NULL, facety = NULL, xlabs = NULL, ylabs = NULL,
  setTheme = TRUE)
}
\arguments{
\item{dataset}{dataframe; a dataframe of result of RunPoolEffect and multiVarRRTab}

\item{manualcolors}{character; the ponit colors.}

\item{manualshapes}{character; the point shape.}

\item{Logscale}{logical; log2 for x-axis (default) or not?}

\item{x}{character; the header name in the dataframe for map to axes of point.}

\item{y}{character; the header name in the dataframe for map to ayes of point.}

\item{lower}{character; the header name in the dataframe for map to lower of the errorbar.}

\item{upper}{character; the header name in the dataframe for map to upper of the errorbar.}

\item{pointsize}{numeric; the point size, default is 3.0 .}

\item{linesize}{numeric; the errorbar line size, default is 0.4 .}

\item{errorbarhheight}{numeric; the errorbar height size, default is 0.2 .}

\item{colorVar}{character; the header name in the dataframe for map to the color of point.}

\item{shapeVar}{character; the header name in the dataframe for map to the shape of point.}

\item{facetx}{character; the header name in the dataframe for map to the facet of row, default is NULL.}

\item{facety}{character; the header name in the dataframe for map to the facet of col, default is NULL.}

\item{xlabs}{character; label for the x-axis.}

\item{ylabs}{character; label for the y-axis.}

\item{setTheme}{logical; whether set the default theme.}
}
\value{
Returns a ggplot object.
}
\description{
Polt a forest with the result of RunPoolEffect and multiVarRRTab base-on the \code{\link[ggplot2]{ggplot}}.
}
\details{
TODO
}
\examples{

library("MetaMicrobiome")
data <- system.file("data", package="MetaMicrobiome", "ggforestDemo.rda")
load(data)
head(ggforestDemoData)
ggforestDemoData$study <- factor(ggforestDemoData$study, 
                               levels=rev(unique(ggforestDemoData$study)))
print(levels(ggforestDemoData$study))
pointcolors <- rev(c("#E41A1C",
                     "#4DAF4A", 
		    "#984EA3", 
		    "#FF7F00", 
		    "#FFFF33", 
		    "#A65628",
		    "#F781BF",
		    "#999999"))
pointshape <- c(18, 20)
data1 <- ggforestDemoData[ggforestDemoData$measure=="Shannon",]
head(data1)
data2 <- ggforestDemoData
head(data2)
p1 <- ggforest(dataset=data1,
      manualcolors=pointcolors,
      manualshapes=pointshape,
      x="est",
      y="study",
      Logscale=TRUE,
      lower="lower",
      upper="upper",
      colorVar="study",
      shapeVar="unite",
      pointsize=3,
      linesize=0.4,
      errorbarhheight=0.05,
      xlabs="Odds Ratio",
      ylabs="",
      setTheme=TRUE)

p2 <- ggforest(dataset=data2,
         manualcolors=pointcolors,
         manualshapes=pointshape,
         x="est",
         y="study",
         Logscale=TRUE,
         lower="lower",
         upper="upper",
         colorVar="study",
         shapeVar="unite",
         pointsize=3,
         linesize=0.4,
         errorbarhheight=0.05,
         xlabs="Odds Ratio",
         ylabs="",
      facetx="measure",
         facety="group",
         setTheme=TRUE)
}
\author{
Shuangbin Xu
}
