% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_low_vector.R
\name{high_low_vector}
\alias{high_low_vector}
\title{creat a vector with high/low versus the threshold.}
\usage{
high_low_vector(var_of_interest = NULL, dataset = NULL,
  threshold = NULL)
}
\arguments{
\item{var_of_interest}{character, the interesting variable names.}

\item{dataset}{dataframe. a dataframe contain the inteeresting variable.}

\item{threshold}{numeric, the threshold.}
}
\value{
a vector with high/low versus the threshold.
}
\description{
Creat a vector with hith/low versus the threshold, suitable for 
analysis with \code{[metafor]{rma}}.
}
\details{
TODO
}
\examples{

library("MetaMicrobiome")
testfile <- system.file("data", package="MetaMicrobiome", "Baxter_16_alpha_data.csv.gz")
data <- read.csv(testfile,header=TRUE, check.names=FALSE)
thresVetor <- getthresholds(dataset=data, 
		c("Shannon", "Observe", "J"),
		type="median")
highlowVector <- high_low_vector(dataset=data, 
	threshold=thresVetor, 
	var_of_interest="Shannon")
head(highlowVector)

}
\author{
Shuangbin Xu
}
