% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RF_model.R
\name{make_RF_model}
\alias{make_RF_model}
\title{model training}
\usage{
make_RF_model(train_data = NULL, study = NULL, numtree = 500,
  number_try = NULL, numbercv = 10, classvariable = NULL)
}
\arguments{
\item{train_data}{list or dataframe, the dataframe contianed the features or a list 
contianed multi-dataframe with features.}

\item{study}{character,if train_data is a list, we can use `study` extract 
the dataframe, default is NULL.}

\item{numtree}{numeric, the number of trees for randomforest, see \code{[randomForest]} details.}

\item{number_try}{numeric, the number of variables randomly sampled as candidates at each split, 
see \code{[randomForest]} details, default is `round(sqrt(ncol(train_data)))`.}

\item{numbercv}{numeric, the number of cross-validation.}

\item{classvariable}{character, the header name of the train dataframe for the classification.}
}
\value{
a model object, see \code{[randomForest]} details.
}
\description{
build a model with training datasets.
}
\details{
TODO
}
\examples{




}
\author{
Shuangbin Xu
}
