% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_low_vector.R
\name{MultiHighLow}
\alias{MultiHighLow}
\title{creat list of vectors with high/low versus the threshold.}
\usage{
MultiHighLow(var_of_interest = NULL, dataset = NULL,
  threshold = NULL)
}
\arguments{
\item{var_of_interest}{vector, the vector of interesting variables.}

\item{dataset}{dataframe, the dataframe contained the interesting variables.}

\item{threshold}{vector, the threshold values vector.}
}
\value{
list of vector with high/low versus the threshold.
}
\description{
Creat list of vetors with high/low versus the threshold, suitable for 
analysis with \code{[metafor]{rma}}
}
\details{
TODO
}
\examples{

library("MetaMicrobiome")
file <- system.file("data", package="MetaMicrobiome", "Baxter_16_alpha_data.csv")
data <- read.csv(file, header=TRUE, check.names=FALSE)
thresVetor <- getthresholds(dataset=data, 
			c("Shannon", "Observe", "J"), 
			type="median")
multiVariableHL <- MultiHighLow(var_of_interest=c("Shannon", 
						     "Observe",
						     "J"),
				dataset=data, 
				threshold=thresVetor)
head(multiVariableHL)


}
\author{
Shuangbin Xu
}
