% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Risk_measure.R
\name{multiRunRR}
\alias{multiRunRR}
\title{Summary measures base on epiR for multi variables}
\usage{
multiRunRR(multiHighLow = NULL, metadavector = NULL,
  prefix = "Disease", grouptype = "Case", method = "cohort.count",
  conf.level = 0.95, score = "OR.strata.score", ...)
}
\arguments{
\item{multiHighLow}{list, mulit variable high-lower-vector.}

\item{metadavector}{dataframe,  metada dataframe contained the group information.}

\item{prefix}{character, the header names of the metada data frame, default is 'Group'.}

\item{grouptype}{character, the positive group names, default is "Case".}

\item{method}{characer, a character string indicating the study design on which the
tabular data has been based. Options are \code{cohort.count}, \code{cohort.time}, \code{case.control},
or \code{cross.sectional}, default is \code{cohort.count}. See \code{\link[epiR]{epi.2by2}} for details.}

\item{conf.level}{numeric, magnitude of the returned confidence intervals. Must be a
single number between 0 and 1. See \code{\link[epiR]{epi.2by2}} for details.}

\item{score}{character, Wald and score confidence intervals for the effect
value for each strata, default is \code{OR.strata.score}, See \code{\link[epiR]{epi.2by2}} for details .}

\item{...}{Additional arguments passed to \code{\link[epiR]{epi.2by2}}.}
}
\value{
the summary measures of risk and a chi-squared test
}
\description{
Computes summary measures of risk and a chi-squared test for difference
in the observed proportions from count data presented in a 2 by 2 table
with high low vector. With multiple strata the function returns crude
and Mantel-Haenszel adjusted measures of association and chi-squared
tests of homogeneity for multi variables (based on \code{[epiR]{epi.2by2}}.
}
\details{
TODO
}
\examples{



}
\author{
ShuangbinXu
}
