% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Risk_measure.R
\name{multiVarRRTab}
\alias{multiVarRRTab}
\title{collate results for summary measure of multi-variables}
\usage{
multiVarRRTab(multiRunRRTab = NULL, var_of_interest = NULL)
}
\arguments{
\item{multiRunRRTab}{list, the results of the}

\item{var_of_interest}{vector, the interesting variables.}
}
\value{

}
\description{
Collate results for summary measure of multi-variables.
}
\details{
TODO
}
\examples{



}
\author{
Shuangbin Xu
}
