% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_test.R
\name{predict_test}
\alias{predict_test}
\title{Model Predictions}
\usage{
predict_test(model, teststudy, dataset, classvariable, classtype,
  Trainstudy)
}
\arguments{
\item{model}{object, a model object for predict.}

\item{teststudy}{character, the names of the dataset, if the dataset is a list.}

\item{dataset}{list or dataframe, the test dataset.}

\item{classvariable}{character, the header name of the train or test dataframe for the classification.}

\item{classtype}{character, the name of the positive group for classification.}

\item{Trainstudy}{chraracter, the name of the origin data of the model.}
}
\value{
a list contianed the predict prob and roc results with sensitivity and specificity.
}
\description{
predict the test datasets with the results of the models.
}
\details{
TODO
}
\examples{


}
\author{
Shuangbin Xu
}
