% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RF_model.R
\name{predict_train}
\alias{predict_train}
\title{models results}
\usage{
predict_train(model, study, minus = FALSE, classtype)
}
\arguments{
\item{model}{object, a model object.}

\item{study}{character, a names of training datasets.}

\item{minus}{logical, whether minus the part of the `study`, default is FALSE.}

\item{classtype}{character, the name of the positive group for classificaion.}
}
\value{
a dataframe for the roc curve plot.
}
\description{
Predict the training datasets for a model
}
\details{
TODO
}
\examples{



}
\author{
Shuangbin Xu
}
