% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCplot.R
\name{ROCplot}
\alias{ROCplot}
\title{plot the roc curve}
\usage{
ROCplot(rocplotdata, x, y, xlab, ylab, roccolors, legendkeyheight = 0.05,
  legendposition = c(0.67, 0.17), ...)
}
\arguments{
\item{rocplotdata}{dataframe, a dataframe of result contained the sensitivity and specificity.}

\item{x}{character, the header name in dataframe.}

\item{y}{character, the header name in dataframe.}

\item{xlab}{character, the label for the x-axis.}

\item{ylab}{character, the label for the y-axis.}

\item{roccolors}{vector, the colors for the roc curve.}

\item{legendkeyheight}{the height of the legend, default is `unit(0.05, "mm")`.}

\item{legendposition}{vector, the position of legend, default is `c(0.67, 0.17)`.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{aes}}}
}
\value{
Returns a ggplot object.
}
\description{
Plot the ROC curve base on the \code{[ggplot2]}
}
\details{
TO DO
}
\examples{



}
\author{
Shuangbin Xu
}
