% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Risk_measure.R
\name{tidy_data}
\alias{tidy_data}
\title{collate results for summary measure}
\usage{
tidy_data(multiRunRRTab, var_of_int)
}
\arguments{
\item{multiRunRRTab}{list, the results of}

\item{var_of_int}{vector, interesting variables}
}
\value{
a results of dataframe for summary measure.
}
\description{
Collate results for summary measure
}
\details{
TODO
}
\examples{



}
\author{
ShuangbinXu
}
